+++ taiga/base/utils/contenttypes.py
@@ -14,18 +14,10 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-
-from django_jinja import library
-from django_sites import get_by_id as get_site_by_id
-
-from taiga.front.urls import urls
+from django.apps import apps
+from django.contrib.contenttypes.management import update_contenttypes
 
 
-@library.global_function(name="resolve_front_url")
-def resolve(type, *args):
-    site = get_site_by_id("front")
-    url_tmpl = "{scheme}//{domain}{url}"
-
-    scheme = site.scheme and "{0}:".format(site.scheme) or ""
-    url = urls[type].format(*args)
-    return url_tmpl.format(scheme=scheme, domain=site.domain, url=url)
+def update_all_contenttypes(**kwargs):
+    for app_config in apps.get_app_configs():
+        update_contenttypes(app_config, **kwargs)