+++ taiga/projects/management/commands/sample_data.py
@@ -21,7 +21,6 @@ from django.core.management.base import BaseCommand
 from django.db import transaction
 from django.utils.timezone import now
 from django.conf import settings
-from django.contrib.webdesign import lorem_ipsum
 from django.contrib.contenttypes.models import ContentType
 
 from sampledatahelper.helper import SampleDataHelper