+++ taiga/users/migrations/0014_auto_20151005_1357.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import models, migrations
+import django.contrib.auth.models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('users', '0013_auto_20150901_1600'),
+    ]
+
+    operations = [
+        migrations.AlterModelManagers(
+            name='user',
+            managers=[
+                ('objects', django.contrib.auth.models.UserManager()),
+            ],
+        ),
+        migrations.AlterField(
+            model_name='user',
+            name='last_login',
+            field=models.DateTimeField(verbose_name='last login', blank=True, null=True),
+        ),
+        migrations.AlterField(
+            model_name='user',
+            name='new_email',
+            field=models.EmailField(verbose_name='new email address', blank=True, null=True, max_length=254),
+        ),
+    ]